// apps/admin/src/types/adoptions.ts
export type AdoptionStatus = "pending" | "approved" | "rejected";

export interface AdoptionRequestSummary {
  id: string;
  nome_adotante?: string;
  telefone?: string;
  animal_especifico?: string;
  status?: AdoptionStatus;
  submittedAt?: string;
  expiresAt?: string;
}

export interface AdoptionRequest extends AdoptionRequestSummary {
  idade?: string;
  estado_civil?: string;
  profissao?: string;
  empresa?: string;
  endereco?: string;
  email?: string;
  redes_sociais?: string;
  qtd_adultos?: string;
  criancas?: string;
  renda_mensal?: string;
  acordo?: string;
  alergia?: string;
  motivo?: string;
  porte?: string;
  sexo?: string;
  idade_animal?: string;
  personalidade?: string;
  atividade?: string;
  responsavel?: string;
  horas_sozinho?: string;
  passeios?: string;
  tipo_moradia?: string;
  proprietario_permite?: string;
  detalhes_moradia?: string;
  moradores?: string;
  areas_frequentar?: string;
  periodos?: string;
  dormir?: string;
  acesso?: string;
  outros_animais?: string;
  castrados?: string;
  ja_teve?: string;
  destino_antigos?: string;
  veterinario?: string;
  racao?: string;
  coleira?: string;
  ciencia_adaptacao?: string;
  tempo_adaptacao?: string;
  adestrador?: string;
  motivo_nao_adestrar?: string;
  carro?: string;
  financeiro_vet?: string;
  vacinas?: string;
  gasto_mensal?: string;
  divulgacao?: string;
  noticias?: string;
  visitas?: string;
  fotos_adocao?: string;
  contribuicao?: string;
  compromisso_vida?: string;
  gravidez?: string;
  viagem?: string;
  mudanca_menor?: string;
  mudanca_longe?: string;
  separacao?: string;
  falecimento?: string;
  perder?: string;
  doenca?: string;
  morder?: string;
  destruicao?: string;
  xixi_errado?: string;
  enxoval?: string;
  devolucao?: string;
  termo_nao_repassar?: string;
  obs?: string;
}